Solve problem where Intellij was passing test to get resource, but not Gradle
The problem was that the Gradle tests run from a special build file that
Gradle creates.  Regarding the test file that was the desired resourse,
intellij could see that file b/c it was right there in the parent folder,
but Gradle was not copying that file to it's special build folder, so when
running the tests from it's own build folder, it's like the file wasn't there.

This solved the problem - putting the language in the build.gradle folder.
Taken from here: https://stackoverflow.com/questions/27046642/access-file-in-junit-test-in-gradle-environment # Please enter the commit message for your changes. Lines starting # with '#' will be ignored, and an empty message aborts the commit.

task copyTestResources(type: Copy) {
      from "${projectDir}/src/test/java/httpServerTests/httpLogicTests/responsesTests"
          into "${buildDir}/classes/java/test/httpServerTests/httpLogicTests/responsesTests"
}